pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Repository-Mapper"
version = "0.6.1"
description = "Scan repositories for cross-file references and generate dependency graphs"
readme = "README.md"
license = {text = "GPL"}
requires-python = ">=3.9"
authors = [
    {name = "Sam Fleming", email = "sam@samplayskeys.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GPL License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
]
keywords = ["repository", "dependency", "graph", "mermaid", "yaml", "json"]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
toml = [
    "toml>=0.10.2",
]

[project.scripts]
repomap = "cli:main"

[project.urls]
Homepage = "https://github.com/SamPlaysKeys/Repository-Mapper"
Documentation = "https://github.com/SamPlaysKeys/Repository-Mapper#readme"
Repository = "https://github.com/SamPlaysKeys/Repository-Mapper"

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
include = ["graph*", "scanner*", "exporters*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.coverage.run]
source = ["graph", "scanner", "exporters"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
